call brand vendor email using event
